requirements.txt
streamlit
Pillow
python-docx
pikepdf
tempfile